// Tasks/OrchestratorV1/interfaces.ts
export interface IStageApproval {

    status: ri.ApprovalStatus;
    count: number;

}